+++ packages/vite-plugin/test-data/codegen-packages/packages/logging/build.config.mjs
@@ -7,5 +7,6 @@ export default {
                 }
             ]
         }
-    }
+    },
+    servicesModule: "./customServices"
 };